Automate the properties table generation in readme files
Added scripts to automate the process of reading properties from src and
generate readme file table based on that.

There are couple of steps in the script:
1. It tries to read the current data from the readme file and generate a properties.json inside docs/readme directory of each component.
2. It reads the data from js file of the component and generate another properties.json inside docs/src directory of each component.
3. Uses Plop readme action to generate and update the current readme file, from readme template under plop-templates/readme/README.md
4. Within the previous task, in data function, it merge the information from source file and readme to update the readme file

Also added the prerelease script to package.json to run this script before each release. Still need some other script to commit and push README files.

Next we should work on the script that update the description in src files. Right now I have the script but need refactoring and decide where it should be added. Based on the decision, maybe we need another small script to read the descriptions as well.
Another next step would be generating the types file as well.